+++ docs/README.md
@@ -42,6 +42,7 @@
         1. String : ex)백엔드,레벨1,자동차경주
 6. Pair
     1. List<String>
+7. pairMatching
 
 ### view
