+++ src/main/java/pairmatching/model/MatchingHistory.java
@@ -1,4 +1,8 @@
 package pairmatching.model;
 
+import java.util.List;
+import java.util.Map;
+
 public class MatchingHistory {
+	Map<Level, List<Pair>> matchingHistory;
 }
